requirements.txt
openai>=1.0.0
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0
pynput>=1.7.6
pystray>=0.19.5
Pillow>=10.0.0
pyperclip>=1.8.2
keyring>=24.0.0
